Add a descriptive title to each letter
Simply using the date as the title doesn’t work, as there are shared dates, and some letters that miss them completely. We can combine the date (where available) with the salutation and use that as the overall title. This title can then be hidden from the reading flow.

This also styles the datelines, and fixes a few mistakes where sender was used instead of recipient.